requirements.txt
flask
mysql-connector-python 
